refactor: Moved sensor & controller management to package
