+++ pyroengine/__init__.py
@@ -1,2 +1,3 @@
 from .core import *
+from . import engine, sensors
 from .version import __version__
